start converting graph to Rust
Due to the nature of this network architecture and the way the graph must be redrawn between rounds of training, it does not seem possible to combine Rust and Python. Because this means that the entire project will need to be in Rust, the data structures should be rewritten to be usable across the modules.