mask output layer
Our training data specifies the target value for the root node, as this
is the one of primary interest. However, as there is not a convenient
way to select tensor elements by value (i.e. select the root embedding
at the appropriate position from each sample within a batch), we need to
provide complete gold standards (embedding values over all positions) to
the loss function. This means masking the non-root outputs and targeting
0 on those positions for now.